add initial vsock support
This adds the ability to create and use vsock sockets, which can serve
as a communication mechanism between multiple VMs or VM and host.

VSOCK are not available everywhere, so this feature is guarded by the
USE_VSOCK flags and test to check vsock functionality requires the
local CID feature, available on Linux 5.6+ with the vsock_loopback
driver.

Signed-off-by: Petre Eftime <epetre@amazon.com>