s2n: enable OCSP stapling if supported
OCSP stapling is not available in all configurations, for example
BoringSSL doesn't currently support it. To allow BoringSSL to work as
a LibCrypto backend for s2n, aws-c-io should first check if it can be enabled
and then ignore it if can't.

This fix requires https://github.com/awslabs/s2n/pull/2253 to be
able to function correctly, otherwise it will still behave as is it
did previously.

Signed-off-by: Petre Eftime <epetre@amazon.com>